+++ packages/@zenithui/utils/src/index.tsx
@@ -1,6 +1,14 @@
 export { default as useDeviceType } from "./hooks/use-device"
 export { default as useTheme } from "./hooks/use-theme"
-export { useDebounce, usePrevious } from "./hooks/use-primitives"
+export { default as useGeolocation } from "./hooks/use-geo-location"
+export { default as useHover } from "./hooks/use-hover"
+export {
+  useDebounce,
+  usePrevious,
+  useDefault,
+  useDocumentTitle,
+  useFavicon,
+} from "./hooks/use-primitives"
 export {
   cn,
   sortByKey,