+++ packages/@zenithui/day-picker/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @zenithui/day-picker
 
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @zenithui/utils@0.2.1
+
 ## 1.0.2
 
 ### Patch Changes