+++ packages/@zenithui/time-picker/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @zenithui/time-picker
 
+## 1.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @zenithui/utils@0.2.1
+
 ## 1.1.1
 
 ### Patch Changes